Refactor Event Handlers into separate files, and convert timestamps.
