+++ src/EventHandlers/CLPool.ts
@@ -1,6 +1,3 @@
-/*
- * Please refer to https://docs.envio.dev for a thorough guide on all Envio indexer features
- */
 import {
   CLPool,
   CLPool_Burn,
